db-sync: Refactor Shelley+ related types
