+++ cardano-db-sync/src/Cardano/DbSync/Era/Shelley/Query.hs
@@ -14,8 +14,8 @@ module Cardano.DbSync.Era.Shelley.Query
 
 import qualified Cardano.Crypto.Hash as Crypto
 import           Cardano.Db
+import           Cardano.DbSync.Era.Shelley.Types
 import           Cardano.DbSync.Era.Shelley.Util (unKeyHashRaw)
-import           Cardano.DbSync.Types
 import           Cardano.DbSync.Util
 
 import           Cardano.Slotting.Slot (SlotNo (..))
